requirements.txt
django-webpack-loader==0.6.0
fabric==2.1.3
